requirements.txt
python-dotenv
requests
flask
pyTelegramBotAPI
psycopg2
llama-cpp-python